refactor: abstrai erros da controller e aplica next-connect
